Remove references to blindness
I tweaked the text to avoid referencing blindness, which can be considered insensitive. Hopefully the new version remains fun enough. 

(Also removed the reference to John Connor because I was honestly confused about that one for a good minute-- I enjoy a pop culture reference but this one felt like it was getting in the way of clarity.)